#12 move preview to state class
